requirements.txt
mysql-connector==2.1.6
lxml==4.3.3